completed

+++ src/main/java/com/cydeo/entity/BaseEntity.java
@@ -15,5 +15,5 @@ import java.math.BigInteger;
 public class BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private BigInteger id;
+    private Long id;
 }
